inventory btn ctrl

+++ nuxt/pages/departs.vue
@@ -16,7 +16,7 @@
             row-key
             :data="tableList"
             :row-config="{isHover: true}"
-            :tree-config="{rowField:'id', transform: true, accordion: true, line: true}"
+            :tree-config="{rowField:'id', accordion: true, line: true, expandAll: true}"
             :radio-config="{trigger: 'row', labelField: 'name', highlight: true, range: true, strict: false}"
             @cell-dblclick="cellDBLClickEvent">        
             <vxe-column field="name" title="名称" type="radio" tree-node></vxe-column>
@@ -38,7 +38,7 @@
                     </vxe-form-item>
                     <vxe-form-item field="name" title="名称" :span="12" :item-render="{}">
                         <template #default="{ data }">
-                            <vxe-input v-model="data.name" placeholder="请输入名称"></vxe-input>
+                            <vxe-input v-model="data.name" placeholder="请输入名称" @change="checkName()"></vxe-input>
                         </template>
                     </vxe-form-item>
                     <vxe-form-item field="status" title="状态" :span="12" :item-render="{}">
@@ -74,8 +74,8 @@
                     <vxe-form-item align="center" title-align="left" :span="24">
                         <template #default>
                             <vxe-button type="submit">提交</vxe-button>
-                            <vxe-button @click="banEvent()">{{banBtn}}</vxe-button>
-                            <vxe-button @click="deleteEvent()">删除</vxe-button>
+                            <vxe-button @click="banEvent()" :disabled="isEdit?false:true">{{banBtn}}</vxe-button>
+                            <vxe-button @click="deleteEvent()" :disabled="isEdit?false:true">删除</vxe-button>
                         </template>
                     </vxe-form-item>
                 </vxe-form>
@@ -204,7 +204,7 @@ export default {
         saveEvent() {
             this.submitLoading = true
             var $table = this.$refs.xTable
-            if(this.selectRow && this.isEdit) {      
+            if(this.isEdit) {      
                 this.$nuxt.$emit('isEdit', this.formDataTemp, this.formData, res => {
                     if(!res) {
                         this.$message({ message: '数据未修改, 此次未提交'})
@@ -227,6 +227,9 @@ export default {
                         this.isEdit = false
                         this.$message({ message: '保存成功', type: 'success' })
                         Object.assign(this.selectRow, this.formData)
+                        this.selectRow = null
+                    } else {
+                        this.$message({ message: res.data, type: 'error' })
                     }
                 }).catch(err => {
                     this.submitLoading = false
@@ -255,7 +258,9 @@ export default {
                             })
                         }                        
                         $table.reloadData(this.tableData)
-                    }                    
+                    } else {
+                        this.$message({ message: res.data, type: 'error' })
+                    }                  
                 }).catch(err => {
                     this.submitLoading = false
                     this.$message({ message: err, type: 'error' })
@@ -280,29 +285,57 @@ export default {
             Object.assign(this.selectRow, this.formData)
         },
         deleteEvent() {
-            this.$confirm('此操作将永久删除该记录, 是否继续?', '提示', {
-                confirmButtonText: '确定',
-                cancelButtonText: '取消',
-                type: 'warning'
-            }).then(() => {
-                var k = { id: null }
-                k['id'] = this.formData.id
-                this.$axios({
-                    method: 'DELETE',
-                    url: '/api/departs',
-                    params: k
-                }).then(res => {
-                    this.submitLoading = false
-                    if(res.data=='OK') {
-                        this.showEdit = false
-                        this.$refs.xTable.remove(this.selectRow)
-                        this.$message({ message: '删除成功', type: 'success' })
-                    }
-                }).catch(err => {
-                    this.submitLoading = false
-                    this.$message({ message: err, type: 'error' })
+            if(this.formData.status==1||this.formData.status==0) {
+                this.$confirm('此操作将永久删除该记录, 是否继续?', '提示', {
+                    confirmButtonText: '确定',
+                    cancelButtonText: '取消',
+                    type: 'warning'
+                }).then(() => {
+                    var k = { id: null }
+                    k['id'] = this.formData.id
+                    this.$axios({
+                        method: 'DELETE',
+                        url: '/api/departs',
+                        params: k
+                    }).then(res => {
+                        this.submitLoading = false
+                        if(res.data=='OK') {
+                            this.showEdit = false
+                            this.$refs.xTable.remove(this.selectRow)
+                            this.$message({ message: '删除成功', type: 'success' })
+                        } else {
+                            this.$message({ message: res.data, type: 'error' })
+                        }
+                    }).catch(err => {
+                        this.submitLoading = false
+                        this.$message({ message: err, type: 'error' })
+                    })
                 })
-            })
+            } else {
+                this.$message({ message: '当前记录不可删除', type: 'warning' })
+            }
+        },
+        checkName() {
+            var status = true, data=this.$refs.xTable.getTableData().tableData
+            a:for(var i=0;i<data.length;i++) {
+                if(data[i].name==this.formData.name) {
+                    status = false
+                    break a
+                }
+                if(data[i].children) {
+                    for(var j=0;j<data[i].children.length;j++) {
+                        if(data[i].children[j].name==this.formData.name) {
+                            status = false
+                            break a
+                        }
+                    }
+                }
+            }
+            if(!status) {
+                this.$message({ message: '名称已存在', type: 'error' })
+                this.formData.name = null
+            }
+            return status
         },
         searchEvent () {
             const filterName = XEUtils.toValueString(this.filterName).trim().toLowerCase()
